File moved to Cryptography -> AuthCodes.h
