+++ src/world/Server/Packets/Handlers/CharacterHandler.cpp
@@ -32,7 +32,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Server/Packets/SmsgCharEnum.h"
 #include "Management/Guild/GuildMgr.hpp"
 #include "Server/CharacterErrors.h"
-#include "AuthCodes.h"
+#include "Cryptography/AuthCodes.h"
 #include "Server/Script/ScriptMgr.h"
 #include "Utilities/Strings.hpp"
